schema-based validations pt 1 (#431)
* don't allow schemas to be defined at the service level

* remove some schema-related types that didn't add any value

* export more schema-related types for direct use by modules

* export RequiredProperties attribute of object param schema for direct use by modules

* DRY up object param and array param marshaling

* move root of schema tree from plan to plan properties

* remove unused description field

* clean up top-level schema (ParameterSchema)

* make schemas on plan NOT a pointer

* refactor how modules define plan schemas

* use constant for json schema version

* s/ParameterSchema/PropertySchema/g

* replace simple and numeric schemas with string, int, and float schemas

* add validation capability to schema types

* add default value fields to array and object prop schemas

* adding tests for schema types

* fix lint errors

* add default firewall rules to select schemas

* improve signature for custom property validators

* make property validations invoke custom validation hooks

* drive binding param validation off schema

* drive update param validation off schema

* drive provisioning param validation off schema

* fix lifecycle test failures

* remove support for validating multipleOf for floats

* err... correct name of new unit test

* validate "additional"

* refactoring api/upate

* fix update bug where we don't return early when we should

* fix bug in unit test that popped up because of correction in api logic

* lint/spelling fix

* fix validation of nil params

* fix tag handling-- allow tags param to safely be null

* remove errant debugging statement

* remove accidental change to contrib; fix typo

* fix accidental modification to prov params during update

* add very minor optimization to update

+++ pkg/services/mysql/dbms_types.go
@@ -5,10 +5,10 @@ import "github.com/Azure/open-service-broker-azure/pkg/service"
 type dbmsProvisioningParameters struct {
 	SSLEnforcement   string         `json:"sslEnforcement"`
 	FirewallRules    []firewallRule `json:"firewallRules"`
-	Cores            *int           `json:"cores"`
-	Storage          *int           `json:"storage"`
+	Cores            *int64         `json:"cores"`
+	Storage          *int64         `json:"storage"`
 	HardwareFamily   string         `json:"hardwareFamily"`
-	BackupRetention  *int           `json:"backupRetention"`
+	BackupRetention  *int64         `json:"backupRetention"`
 	BackupRedundancy string         `json:"backupRedundancy"`
 }
